Refactor location groups
